Fix the docs refs and simplify FlowGraph ID input
