+++ flowfile/flowfile/__init__.py
@@ -20,6 +20,14 @@ from flowfile.api import (open_graph_in_editor)
 from flowfile_frame.flow_frame import (
     FlowFrame
 )
+from flowfile_core.schemas.cloud_storage_schemas import FullCloudStorageConnection
+from flowfile_core.flowfile.flow_graph import FlowGraph
+from flowfile_core.flowfile.flow_data_engine.flow_data_engine import FlowDataEngine
+from flowfile_core.flowfile.flow_node.flow_node import FlowNode
+from flowfile_core.flowfile.flow_data_engine.flow_file_column.main import FlowfileColumn
+from flowfile_core.schemas import input_schema
+from flowfile_core.schemas.schemas import FlowSettings, FlowInformation
+from flowfile_core.schemas import transform_schema
 from flowfile_frame import (read_csv, read_parquet, from_dict, concat, scan_csv, scan_parquet,
                             scan_delta,
                             scan_parquet_from_cloud_storage,
@@ -55,13 +63,15 @@ from polars.datatypes import (
 
 __all__ = [
     # Core FlowFrame classes
-    'FlowFrame', 'GroupByFrame',
+    'FlowFrame', 'GroupByFrame', 'FullCloudStorageConnection',
 
     # Main creation functions
     'read_csv', 'read_parquet', 'from_dict', 'concat', 'scan_csv', 'scan_parquet', 'scan_delta',
     'scan_parquet_from_cloud_storage', 'scan_json_from_cloud_storage',
     'scan_csv_from_cloud_storage', 'get_all_available_cloud_storage_connections', 'create_cloud_storage_connection',
     'del_cloud_storage_connection', 'create_cloud_storage_connection_if_not_exists',
+    'FlowGraph', 'FlowDataEngine', 'input_schema', 'FlowSettings', 'transform_schema',
+    'FlowNode', 'FlowfileColumn', 'FlowInformation',
 
     # Expression API
     'col', 'lit', 'column', 'cum_count', 'len',
