+++ flowfile_core/flowfile_core/flowfile/utils.py
@@ -1,11 +1,15 @@
 import os
-import hashlib
 import json
 import shutil
 
 from datetime import datetime, date, time
 from typing import List
 from decimal import Decimal
+import time
+import random
+import uuid
+import socket
+import hashlib
 
 
 def generate_sha256_hash(data: bytes):
@@ -113,3 +117,29 @@ def batch_generator(input_list: List, batch_size: int = 10000):
             yield input_list
             input_list = []
             run = False
+
+
+def create_unique_id() -> int:
+    """
+    Create a unique id for the flowfile with increased randomness while maintaining 32-bit compatibility
+    Returns:
+        int: unique id within 32 bits (4 bytes)
+    """
+    # Get various entropy sources
+    time_ms = int(time.time() * 1000)
+    pid = os.getpid()
+    random_bytes = random.getrandbits(32)
+    mac_addr = uuid.getnode()  # MAC address as integer
+    hostname = socket.gethostname()
+
+    # Combine all sources into a string
+    seed = f"{time_ms}-{pid}-{random_bytes}-{mac_addr}-{hostname}-{uuid.uuid4()}"
+
+    # Create a hash of all entropy sources
+    hash_obj = hashlib.md5(seed.encode())
+    hash_int = int(hash_obj.hexdigest(), 16)
+
+    # Ensure the result fits within 32 bits (4 bytes)
+    unique_id = hash_int & 0xFFFFFFFF
+
+    return unique_id
