+++ flowfile_frame/flowfile_frame/utils.py
@@ -95,7 +95,7 @@ def create_flow_graph() -> FlowGraph:
         name=f"Flow_{flow_id}",
         path=f"flow_{flow_id}"
     )
-    flow_graph = FlowGraph(flow_id=flow_id, flow_settings=flow_settings)
+    flow_graph = FlowGraph(flow_settings=flow_settings)
     flow_graph.flow_settings.execution_location = 'local'  # always create a local frame so that the run time does not attempt to use the flowfile_worker process
     return flow_graph
